+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.9.3] - 2020-07-13
 ### Added
 - Crafting a repair kit with a stick will empty it completely (materials cannot be returned)
 - JEI support for salvager
+### Changed
+- Reworked advancements [#151]
 ### Fixed
 - Crash on server startup (KeyTracker and KeyBinding)
 - Materials not displaying their grade